+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## 0.1.6
+ - Support for Themes (dropdown list)
+ - Initial themes: Default, Timber.
+
 ## 0.1.5
  - Fix for preferredSort (became broken in v0.1.4)
  - Modified plugin's internal directory structure